refactor(consensus): make the Raft nameservice node embeddable
Everything a process needs to run a Raft-replicated nameservice node
already existed — but half of it lived in the fluree-db-server binary,
and the other half was ~180 lines of assembly inside
`FlureeServerBuilder::build`. An embedding process had to re-derive
both from the server's source.

`fluree-db-server/src/raft.rs` moves to
`fluree-db-consensus::raft::integration` verbatim. It never referenced
a server type — the one `crate::` in the file was inside its own tests
— so the move is a relocation, not a redesign. The server re-exports it
at the old path.

The assembly becomes `raft::embedded::EmbeddedRaftNode::attach`: the
queued committer, the ledger-cache watermark hookup, the worker
supervisor, the leader watcher (evictor, liveness monitor, plus
whatever the host adds), and the release task — owned together so they
shut down in the order that keeps the content store consistent:
workers, then leader tasks, then the Raft core, then the release drain.
The background indexer stays a host contribution because this crate
does not depend on fluree-db-indexer; omit it and nothing indexes.

`Fluree::default_index_config` is now public, so an embedder hands the
commit workers the engine's own thresholds instead of constructing
them independently and letting staging and novelty backpressure drift.

The server itself now builds through `EmbeddedRaftNode`, which is how
this is validated: the nine-test cluster suite — failover, follower
forwarding, read-your-write across nodes, liveness demotion, concurrent
writes — runs through exactly the code an embedder calls. A new
`tests/it_embedded_node.rs` stands a node up with no server dependency
at all, mounts the routers at a host-chosen prefix, writes through the
committer, and checks the Raft log advanced and the replicated head and
the engine's cache both agree on the committed t.
