+++ docs/design/raft-command-queue.md
@@ -23,6 +23,8 @@ The Raft consensus crate (`fluree-db-consensus/src/raft/`) is structured as a se
 | `NameServiceApp`           | `app.rs`                                                 | The pure reduction: routes a `Command` through `state_machine::apply`, mirrors membership, owns the snapshot codec. |
 | `NameServiceObserver`      | `app.rs`                                                 | The effects: event bus, waiter resolution, staged receipts, releases, ledger-cache watermark. Captured under the state lock, published after it drops. |
 | `StateMachineAdapter`      | `state_machine_adapter.rs`                               | The two above composed with `fluree_raft_core::state_machine::StateMachineAdapter`, which owns last-applied, membership, and snapshot persistence. |
+| `RaftIntegration`          | `integration.rs`                                         | One-call consensus assembly: storage, adapters, `Raft`, `RaftNameService`, routers, channels. Host-agnostic — moved here from the server binary, which it never depended on. |
+| `EmbeddedRaftNode`         | `embedded.rs`                                            | Wires a `Fluree` engine to a `RaftIntegration`: the committer, worker supervisor, leader watcher, release task, and their shutdown order. What an embedding process calls instead of re-deriving `FlureeServerBuilder::build`. |
 | `LogStore`, `SnapshotStore`| `fluree-raft-core` `log_adapter.rs`, `storage/{fs,memory}.rs` | openraft's `RaftLogStorage`/`RaftLogReader`/`RaftSnapshotBuilder`. Local-disk persistence. The backends are openraft-free — they store opaque `Vec<u8>` payloads. |
 | `HttpRaftNetworkFactory`   | `fluree-raft-core` `network.rs`                          | Inter-node RPC (`/raft/vote`, `/raft/append-entries`, `/raft/install-snapshot`) over HTTP. Generic over the type config; `network.rs` here keeps only the nameservice-specific `NetworkConfig` fields. |
 | `RaftAdmin` / `/cluster/*` | `fluree-raft-core` `admin.rs`                            | Operator-facing membership endpoints (`initialize`, `add-learner`, `change-membership`, `status`).         |
