+++ fluree-db-consensus/src/raft.rs
@@ -97,7 +97,9 @@
 
 pub mod app;
 pub mod commit_worker;
+pub mod embedded;
 pub mod eviction_scheduler;
+pub mod integration;
 pub mod liveness_monitor;
 pub mod nameservice;
 pub mod network;
